[REF] removed api-service into shared folder and refactor faculty component
