+++ src/app/admin/speciality/dialog-form/dialog-form.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Inject } from '@angular/core';
-import { ApiService } from '../api.service';
+import { ApiService } from '../../../shared/services/api.service';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { Speciality } from '.././../entity.interface';
