+++ src/app/shared/services/api.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
+import {Observable} from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -11,19 +12,19 @@ export class ApiService {
 
   constructor(private http: HttpClient) { }
 
-  getEntity(entity, action) {
+  getEntity(entity, action): Observable<any> {
     const url = `${this.apiURI}/${entity}/${action}`;
     return this.http.get(url);
   }
-  postEntity(entity, action, body) {
+  postEntity(entity, action, body): Observable<any> {
     const url = `${this.apiURI}/${entity}/${action}`;
     return this.http.post(url, body);
   }
-  delEntity(entity, action, id) {
+  delEntity(entity, action, id): Observable<any> {
     const url = `${this.apiURI}/${entity}/${action}/${id}`;
     return this.http.get(url);
   }
-  updEntity(entity, action, body, id) {
+  updEntity(entity, action, body, id): Observable<any> {
     const url = `${this.apiURI}/${entity}/${action}/${id}`;
     return this.http.post(url, body);
   }
